Make FallbackLoaderTest more robust.
